Further externalizations and improvements!
Transcript is now externalized. It inserts commas between messages that
it recites, and records ACTION messages. Topic is externalized, and a
fatal bug in !topic (with no parameters, to request what the topic is)
is now fixed. Vote (with no parameters) now accepts a trailing space.
Module initialization has been moved to later in the program, allowing
modules to access the IRC connection.